functionality changes

+++ app/code/Qdos/QdosSync/Cron/SyncDeleteProduct.php
@@ -5,15 +5,18 @@
 namespace Qdos\QdosSync\Cron;
 use \Psr\Log\LoggerInterface;
 use \Qdos\QdosSync\Helper\Product;
+use \Qdos\QdosSync\Helper\Data;
 
 class SyncDeleteProduct
 {
     protected $_logger;
     protected $_helperData;
     public function __construct(LoggerInterface $logger,
-                                Product $productHelper){
+                                Product $productHelper,
+                                Data $dataHelper){
         $this->_logger = $logger;
         $this->_helperProduct = $productHelper;
+        $this->_dataHelper = $dataHelper;
     }
 
     public function execute()
@@ -42,7 +45,22 @@ class SyncDeleteProduct
                 date_default_timezone_set('Asia/Kolkata');
                 $this->_resourceConfig->saveConfig('qdosConfig/cron_status/current_cron_updated_time', date("Y-m-d H:i:s"), 'default', 0);
 
-                $syncStatus = $this->_helperProduct->deleteProducts();
+                
+                $syncPermissions = $this->_dataHelper->getSyncPermission(0);
+                if (in_array('Delete Product', $syncPermissions)) {
+                  //Mage::log('category sync executing for store id 0: ', null, 'storesync.log');
+                  $syncStatus = $this->_helperProduct->deleteProducts();
+                }
+                $storeManager = $objectManager->create("\Magento\Store\Model\StoreManagerInterface");
+                $allStores = $storeManager->getStores(true, false);
+                foreach ($allStores as $storeId => $val)
+                {
+                    $syncPermissions = $this->_dataHelper->getSyncPermission($storeId);
+                    if (in_array('Delete Product', $syncPermissions)) {
+                      //Mage::log('category sync executing for store id '.$storeId, null, 'storesync.log');
+                      $syncStatus = $this->_helperProduct->deleteProducts($storeId);
+                    }                  
+                }
 
                 if ($syncStatus == 'success') {
                     $logMsg = 'Qdos Delete Product Sync Successful';
